requirements.txt
motor==3.1.2
aiogram==2.25.1
python-dotenv==1.0.0
pydantic==1.10.7